Use mixin for common logic with PaymentPlan export and PaymentPlanGroup logic
